[+] Added DetachedModuleSpawner and Spawn on NetModule C_DetachModuleObserver
Ive implemented a DetachedModuleSpawner as SceneSingleton
Thus its possible to spawn a detachedModule by providing the ModuleID, Transform and optionally a 2d detach velocity.

The Spawner takes the setup visualPrefab from the ModuleData to instantiate a visual gameobject inside a detachedModulePrefab.
The DetachedModule has a visual representation of the module and applies the provided velocity with a dampening factor.
If the velocity is below a threshold, make the module static.